Use Intl short units for CV durations (3 yrs 9 mths) (#84)
Switch the date-range duration from Intl's long unit form to short —
"3 yrs 9 mths" (en), "3 J, 9 Mon." (de), "3 a 9 m." (es), etc. The Liquid
en fallback now renders the matching "yr/yrs, mth/mths" so it agrees with
the JS output.

+++ assets/js/cv/DateRange.js
@@ -41,8 +41,8 @@ function endpointLabel(locale, point) {
 }
 
 /**
- * A localised, long-form duration for a whole number of months, e.g.
- * "3 years 9 months" / "3 años 9 meses" / "6 years". Intl owns the unit words,
+ * A localised, short-form duration for a whole number of months, e.g.
+ * "3 yrs 9 mths" / "3 a 9 m." / "6 yrs". Intl owns the unit abbreviations,
  * plurals and joining.
  *
  * @param {String} locale
@@ -56,10 +56,10 @@ function durationText(locale, totalMonths) {
 
     const parts = [];
     if (years) {
-        parts.push(new Intl.NumberFormat(locale, { style: 'unit', unit: 'year', unitDisplay: 'long' }).format(years));
+        parts.push(new Intl.NumberFormat(locale, { style: 'unit', unit: 'year', unitDisplay: 'short' }).format(years));
     }
     if (months) {
-        parts.push(new Intl.NumberFormat(locale, { style: 'unit', unit: 'month', unitDisplay: 'long' }).format(months));
+        parts.push(new Intl.NumberFormat(locale, { style: 'unit', unit: 'month', unitDisplay: 'short' }).format(months));
     }
 
     return new Intl.ListFormat(locale, { style: 'narrow', type: 'unit' }).format(parts);
